# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DevOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /note-folders:
    get:
      operationId: listNoteFolders
      tags: [notes]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/NoteFolder" }
    post:
      operationId: createNoteFolder
      tags: [notes]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteFolderInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteFolder" }
  /note-folders/{id}:
    patch:
      operationId: updateNoteFolder
      tags: [notes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteFolderUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteFolder" }
    delete:
      operationId: deleteNoteFolder
      tags: [notes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: No Content
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      parameters:
        - name: folderId
          in: query
          schema: { type: integer }
        - name: pinned
          in: query
          schema: { type: boolean }
        - name: archived
          in: query
          schema: { type: boolean }
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Note" }
    post:
      operationId: createNote
      tags: [notes]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }
    patch:
      operationId: updateNote
      tags: [notes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }
    delete:
      operationId: deleteNote
      tags: [notes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: No Content
  /notes/{id}/duplicate:
    post:
      operationId: duplicateNote
      tags: [notes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }
  /notes/{id}/versions:
    get:
      operationId: listNoteVersions
      tags: [notes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/NoteVersion" }

  # ===================== CODING PROFILES =====================
...
components:
  schemas:
...
    NoteFolder:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        parentId: { type: ["integer", "null"] }
        name: { type: string }
        createdAt: { type: string, format: date-time }
    NoteFolderInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        parentId: { type: integer }
    NoteFolderUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        parentId: { type: ["integer", "null"] }
    Note:
      type: object
      required: [id, title, contentMarkdown, pinned, archived, tags, createdAt, updatedAt]
      properties:
        id: { type: integer }
        folderId: { type: ["integer", "null"] }
        title: { type: string }
        contentMarkdown: { type: string }
        tags:
          type: array
          items: { type: string }
        pinned: { type: boolean }
        archived: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    NoteInput:
      type: object
      required: [title]
      properties:
        folderId: { type: integer }
        title: { type: string, minLength: 1 }
        contentMarkdown: { type: string }
        tags:
          type: array
          items: { type: string }
    NoteUpdate:
      type: object
      properties:
        folderId: { type: ["integer", "null"] }
        title: { type: string, minLength: 1 }
        contentMarkdown: { type: string }
        tags:
          type: array
          items: { type: string }
        pinned: { type: boolean }
        archived: { type: boolean }
    NoteVersion:
      type: object
      required: [id, noteId, contentMarkdown, createdAt]
      properties:
        id: { type: integer }
        noteId: { type: integer }
        contentMarkdown: { type: string }
        createdAt: { type: string, format: date-time }

    # ---------- CODING PROFILES ----------
...